+++ Code/src/main/java/com/huaDevelopers/dao/Interfaces/CustomerDAO.java
@@ -14,6 +14,6 @@ public interface CustomerDAO {
 
 	public List<Customer> listAllCustomers();
 
-	public void removeCustomer(int id);
+	public void removeCustomer(Long id);
 
 }
