+++ Code/src/main/java/com/huaDevelopers/data/Services/Interfaces/CustomerService.java
@@ -14,5 +14,5 @@ public interface CustomerService {
 
 	public List<Customer> listAllCustomers();
 
-	public void removeCustomer(int id);
+	public void removeCustomer(Long id);
 }
